+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Group changes to describe their impact on the project, as follows:
 - Unread messages indicator in chat conversation that separates read & unread messages
 - Notify incoming/outgoing calls on bluetooth devices using self-managed connections from telecom manager API (disables SDK audio focus)
 - Ask Android to not process what user types in an encrypted chat room to improve privacy, see [IME_FLAG_NO_PERSONALIZED_LEARNING](https://developer.android.com/reference/android/view/inputmethod/EditorInfo#IME_FLAG_NO_PERSONALIZED_LEARNING)
+- SIP URIs in chat messages are clickable to easily initiate a call
 - New video call UI on foldable device like Galaxy Z Fold
 - Setting to automatically record all calls
 - When using a physical keyboard, use left control + enter keys to send message